fix: boosts not being able to be stored into the database
The column wasn't auto_increment since some time in the past
because of compatibility and stability issues we had back then
in the Core's database abstraction.
Somehow the broken code got published and nobody realized for some time.

In addition, H2 does not *really* support `RETURNING`
but JOOQL can still support it, if we tell it that the id column
we want is actually the primary key.